+++ MinimedKitUI/MinimedPumpSettingsViewController.swift
@@ -8,12 +8,60 @@
 import UIKit
 import LoopKitUI
 import MinimedKit
+import RileyLinkBLEKit
 import RileyLinkKitUI
 import LoopKit
 
 class MinimedPumpSettingsViewController: RileyLinkSettingsViewController {
 
     let pumpManager: MinimedPumpManager
+
+    private var ops: PumpOps {
+        return pumpManager.pumpOps
+    }
+
+    // MARK: - Formatters
+
+    private lazy var dateFormatter: DateFormatter = {
+        let dateFormatter = DateFormatter()
+
+        dateFormatter.dateStyle = .none
+        dateFormatter.timeStyle = .medium
+
+        return dateFormatter
+    }()
+
+    private lazy var measurementFormatter: MeasurementFormatter = {
+        let formatter = MeasurementFormatter()
+
+        formatter.numberFormatter = decimalFormatter
+
+        return formatter
+    }()
+
+    private lazy var decimalFormatter: NumberFormatter = {
+        let decimalFormatter = NumberFormatter()
+
+        decimalFormatter.numberStyle = .decimal
+        decimalFormatter.minimumSignificantDigits = 5
+
+        return decimalFormatter
+    }()
+
+    private lazy var integerFormatter = NumberFormatter()
+
+    private func cellForRow(_ row: CommandsRow) -> UITableViewCell? {
+        return tableView.cellForRow(at: IndexPath(row: row.rawValue, section: Section.commands.rawValue))
+    }
+
+    private var pumpState: PumpState? {
+        didSet {
+            if let cell = cellForRow(.tune) {
+                cell.setTuneInfo(lastValidFrequency: pumpState?.lastValidFrequency, lastTuned: pumpState?.lastTuned, measurementFormatter: measurementFormatter, dateFormatter: dateFormatter)
+            }
+        }
+    }
+
     
     init(pumpManager: MinimedPumpManager) {
         self.pumpManager = pumpManager
@@ -82,6 +130,7 @@ class MinimedPumpSettingsViewController: RileyLinkSettingsViewController {
         case actions
         case settings
         case rileyLinks
+        case commands
         case delete
     }
 
@@ -104,6 +153,18 @@ class MinimedPumpSettingsViewController: RileyLinkSettingsViewController {
         case useMySentry
     }
 
+    private enum CommandsRow: Int, CaseIterable {
+        case tune
+        case mySentryPair
+        case dumpHistory
+        case fetchGlucose
+        case getPumpModel
+        case pressDownButton
+        case readPumpStatus
+        case readBasalSchedule
+    }
+
+
     // MARK: UITableViewDataSource
 
     override func numberOfSections(in tableView: UITableView) -> Int {
@@ -117,9 +178,12 @@ class MinimedPumpSettingsViewController: RileyLinkSettingsViewController {
         case .actions:
             return ActionsRow.allCases.count
         case .settings:
-            return SettingsRow.allCases.count
+            let settingsRowCount = pumpManager.state.pumpModel.hasMySentry ? SettingsRow.allCases.count : SettingsRow.allCases.count - 1
+            return settingsRowCount
         case .rileyLinks:
             return super.tableView(tableView, numberOfRowsInSection: section)
+        case .commands:
+            return CommandsRow.allCases.count
         case .delete:
             return 1
         }
@@ -128,9 +192,11 @@ class MinimedPumpSettingsViewController: RileyLinkSettingsViewController {
     override func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
         switch Section(rawValue: section)! {
         case .settings:
-            return LocalizedString("Configuration", comment: "The title of the configuration section in settings")
+            return LocalizedString("Configuration", comment: "The title of the configuration section in MinimedPumpManager settings")
         case .rileyLinks:
             return super.tableView(tableView, titleForHeaderInSection: section)
+        case .commands:
+            return LocalizedString("Commands", comment: "The title of the commands section in MinimedPumpManager settings")
         case .delete:
             return " "  // Use an empty string for more dramatic spacing
         case .info, .actions:
@@ -142,7 +208,7 @@ class MinimedPumpSettingsViewController: RileyLinkSettingsViewController {
         switch Section(rawValue: section)! {
         case .rileyLinks:
             return super.tableView(tableView, viewForHeaderInSection: section)
-        case .info, .settings, .delete, .actions:
+        case .info, .settings, .delete, .actions, .commands:
             return nil
         }
     }
@@ -210,6 +276,36 @@ class MinimedPumpSettingsViewController: RileyLinkSettingsViewController {
             return cell
         case .rileyLinks:
             return super.tableView(tableView, cellForRowAt: indexPath)
+        case .commands:
+            let cell = tableView.dequeueReusableCell(withIdentifier: SettingsTableViewCell.className, for: indexPath)
+
+            switch CommandsRow(rawValue: indexPath.row)! {
+            case .tune:
+                cell.setTuneInfo(lastValidFrequency: pumpState?.lastValidFrequency, lastTuned: pumpState?.lastTuned, measurementFormatter: measurementFormatter, dateFormatter: dateFormatter)
+            case .mySentryPair:
+                cell.textLabel?.text = LocalizedString("MySentry Pair", comment: "The title of the command to pair with mysentry")
+
+            case .dumpHistory:
+                cell.textLabel?.text = LocalizedString("Fetch Recent History", comment: "The title of the command to fetch recent history")
+
+            case .fetchGlucose:
+                cell.textLabel?.text = LocalizedString("Fetch Enlite Glucose", comment: "The title of the command to fetch recent glucose")
+
+            case .getPumpModel:
+                cell.textLabel?.text = LocalizedString("Get Pump Model", comment: "The title of the command to get pump model")
+
+            case .pressDownButton:
+                cell.textLabel?.text = LocalizedString("Send Button Press", comment: "The title of the command to send a button press")
+
+            case .readPumpStatus:
+                cell.textLabel?.text = LocalizedString("Read Pump Status", comment: "The title of the command to read pump status")
+
+            case .readBasalSchedule:
+                cell.textLabel?.text = LocalizedString("Read Basal Schedule", comment: "The title of the command to read basal schedule")
+
+            }
+            return cell
+
         case .delete:
             let cell = tableView.dequeueReusableCell(withIdentifier: TextButtonTableViewCell.className, for: indexPath) as! TextButtonTableViewCell
 
@@ -225,7 +321,7 @@ class MinimedPumpSettingsViewController: RileyLinkSettingsViewController {
         switch Section(rawValue: indexPath.section)! {
         case .info:
             return false
-        case .actions, .settings, .rileyLinks, .delete:
+        case .actions, .settings, .rileyLinks, .delete, .commands:
             return true
         }
     }
@@ -284,6 +380,17 @@ class MinimedPumpSettingsViewController: RileyLinkSettingsViewController {
             )
 
             self.show(vc, sender: sender)
+        case .commands:
+            pumpManager.rileyLinkDeviceProvider.firstConnectedDevice { device in
+                DispatchQueue.main.async {
+                    if let device = device,
+                       let cell = tableView.cellForRow(at: indexPath),
+                       let title = cell.textLabel?.text
+                    {
+                        self.runCommand(CommandsRow(rawValue: indexPath.row)!, device: device, title: title)
+                    }
+                }
+            }
         case .delete:
             let confirmVC = UIAlertController(pumpDeletionHandler: {
                 self.pumpManager.notifyDelegateOfDeactivation {
@@ -305,14 +412,48 @@ class MinimedPumpSettingsViewController: RileyLinkSettingsViewController {
             switch SettingsRow(rawValue: indexPath.row)! {
             case .timeZoneOffset:
                 tableView.reloadRows(at: [indexPath], with: .fade)
-            case .batteryChemistry, .preferredInsulinDataSource, .useMySentry:
+            case .batteryChemistry:
+                break
+            case .preferredInsulinDataSource:
+                break
+            case .useMySentry:
                 break
             }
-        case .info, .actions, .rileyLinks, .delete:
+        case .info, .actions, .rileyLinks, .delete, .commands:
             break
         }
 
         return indexPath
+    }
+
+    private func runCommand(_ command: CommandsRow, device: RileyLinkDevice, title: String) {
+        var vc: CommandResponseViewController?
+
+        switch command {
+        case .tune:
+            vc = .tuneRadio(ops: ops, device: device, measurementFormatter: measurementFormatter)
+        case .mySentryPair:
+            vc = .mySentryPair(ops: ops, device: device)
+        case .dumpHistory:
+            vc = .dumpHistory(ops: ops, device: device)
+        case .fetchGlucose:
+            vc = .fetchGlucose(ops: ops, device: device)
+        case .getPumpModel:
+            vc = .getPumpModel(ops: ops, device: device)
+        case .pressDownButton:
+            vc = .pressDownButton(ops: ops, device: device)
+        case .readPumpStatus:
+            vc = .readPumpStatus(ops: ops, device: device, measurementFormatter: measurementFormatter)
+        case .readBasalSchedule:
+            vc = .readBasalSchedule(ops: ops, device: device, integerFormatter: integerFormatter)
+        }
+
+        vc?.title = title
+
+        if let vc = vc {
+            show(vc, sender: nil)
+        }
+
     }
 }
 
@@ -379,8 +520,10 @@ extension MinimedPumpSettingsViewController: RadioSelectionTableViewControllerDe
 extension MinimedPumpSettingsViewController: PumpManagerStatusObserver {
     public func pumpManager(_ pumpManager: PumpManager, didUpdate status: PumpManagerStatus, oldStatus: PumpManagerStatus) {
         dispatchPrecondition(condition: .onQueue(.main))
-        let suspendResumeTableViewCell = self.tableView?.cellForRow(at: IndexPath(row: ActionsRow.suspendResume.rawValue, section: Section.actions.rawValue)) as! SuspendResumeTableViewCell
-        suspendResumeTableViewCell.basalDeliveryState = status.basalDeliveryState
+        if let suspendResumeTableViewCell = self.tableView?.cellForRow(at: IndexPath(row: ActionsRow.suspendResume.rawValue, section: Section.actions.rawValue)) as? SuspendResumeTableViewCell
+        {
+            suspendResumeTableViewCell.basalDeliveryState = status.basalDeliveryState
+        }
     }
 }
 
@@ -402,5 +545,25 @@ private extension UIAlertController {
 
         let cancel = LocalizedString("Cancel", comment: "The title of the cancel action in an action sheet")
         addAction(UIAlertAction(title: cancel, style: .cancel, handler: nil))
+    }
+}
+
+private extension UITableViewCell {
+
+    func setDetailDate(_ date: Date?, formatter: DateFormatter) {
+        if let date = date {
+            detailTextLabel?.text = formatter.string(from: date)
+        } else {
+            detailTextLabel?.text = "-"
+        }
+    }
+
+    func setTuneInfo(lastValidFrequency: Measurement<UnitFrequency>?, lastTuned: Date?, measurementFormatter: MeasurementFormatter, dateFormatter: DateFormatter) {
+        if let frequency = lastValidFrequency, let date = lastTuned {
+            textLabel?.text = measurementFormatter.string(from: frequency)
+            setDetailDate(date, formatter: dateFormatter)
+        } else {
+            textLabel?.text = LocalizedString("Tune Radio Frequency", comment: "The title of the command to re-tune the radio")
+        }
     }
 }
